ir: Avoid signed overflow in MUL/DIV(NEG, C_I) folding
The rules MUL(NEG(x), c) and DIV(NEG(x), c) negate the constant on a
signed int64_t, which is undefined behavior for INT64_MIN. Negate
through uint64_t instead. The result is unchanged on a release build
since the negation of INT64_MIN wraps to itself.

+++ ir_fold.h
@@ -2216,7 +2216,7 @@ IR_FOLD(DIV(NEG, C_I32))
 IR_FOLD(DIV(NEG, C_I64))
 {
 	op1 = op1_insn->op1;
-	val.i64 = -op2_insn->val.i64;
+	val.i64 = -(uint64_t)op2_insn->val.i64;
 	op2 = ir_const(ctx, val, op2_insn->type);
 	IR_FOLD_RESTART;
 }
